requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
openai==0.27.2
python-dotenv
pytest
black
ruff